requirements.txt
camel-ai[all]==0.2.32
chunkr-ai>=0.0.41
docx2markdown>=0.1.1
gradio>=3.50.2
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.2
python-dotenv==1.0.0
python-multipart==0.0.6
asyncio==3.4.3
aiohttp==3.8.5
